pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ogmem"
version = "0.1.0"
description = "Verifiable, encrypted, user-owned agent memory on 0G Labs"
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
dependencies = [
    "web3>=6.0.0",
    "requests>=2.31.0",
    "openai>=1.0.0",
    "cryptography>=42.0.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=2.7.0",
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.4.0",
]
tui = [
    "textual>=0.60.0",
]
bot = [
    "python-telegram-bot>=21.0",
]
langchain = [
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["ogmem"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"
